Split flat_binary_tree in own header
